% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstFun.R
\name{iter_reconst_fun}
\alias{iter_reconst_fun}
\title{Iterative Reconstruction Algoritm}
\usage{
iter_reconst_fun(cov_la_mat, domain_grid, Y_cent_sm_i, U_sm_i, K,
  fraction = 0.15, max_rep = 5, pre_smooth = FALSE, messages = FALSE)
}
\arguments{
\item{cov_la_mat}{Discretized covariance function over \eqn{[a,b]\times[a,b]}{[a,b]x[a,b]}}

\item{domain_grid}{Equidistant discretization grid in \eqn{[a,b]}{[a,b]}}

\item{Y_cent_sm_i}{Centered function values of the ith function: \eqn{Y_{ij}-\hat\mu(U_{ij}), j=1,\dots,m}{Y_{ij}-\hat\mu(U_{ij}), j=1,...,m},}

\item{U_sm_i}{Discretization points of the ith function: \eqn{U_{i1},\dots,U_{im}}{U_{i1},...,U_{im}}}

\item{K}{Truncation parameter}

\item{fraction}{A value between 0 and 1. (See details)}

\item{max_rep}{Maximal number of iterations (default max_rep=5)}

\item{pre_smooth}{If pre_smooth==TRUE:  Pre-smoothing of the 'observed' part.  (Reconstruction operator: \eqn{L^*}{L*}). If pre_smooth==FALSE (default): FPCA-estimation of the 'observed' part (Reconstruction operator: \eqn{L}{L})}

\item{messages}{Printing messages if the algorithm stops (default messages=FALSE)}
}
\description{
This function iteratively applies the function reconst_fun() in order to reconstruct the missing parts of a function given the observed parts. 
The iterative procedure allows to reconstruct functions when their covariance function cannot be estimated over the total domain. However, the covariance function must be estimated over a sufficiently large part of the domain.
}
\details{
Idea of the procedure: In each iteration the observed function is devided in a upper and lower fragment which are used to reconstruct further upper and lower missing parts. 

The lengths of the upper and lower fragments are determined by the argument 'fraction'. 
Large values of 'fraction' lead to large (i.e., more informative) fragments which can improve the reconstructions, but will not allow to reconstruct large missing parts. 
Small values of 'fraction' lead to small (i.e., less informative) fragments which can worsen the reconstructions, but will allow to reconstruct large missing parts.
}
