% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstFun.R
\name{reconst_fun}
\alias{reconst_fun}
\title{Reconstruct a single partially observed centered function}
\usage{
reconst_fun(cov_la_mat, workGrid, Y_cent_sm_i, U_sm_i, K, pre_smooth = FALSE)
}
\arguments{
\item{cov_la_mat}{Discretized covariance function over \eqn{[a,b]\times[a,b]}{[a,b]x[a,b]}}

\item{workGrid}{Equidistant discretization grid in \eqn{[a,b]}{[a,b]}}

\item{Y_cent_sm_i}{Centered function values of the ith function: \eqn{Y_{ij}-\hat\mu(U_{ij}), j=1,\dots,m}{Y_{ij}-\hat\mu(U_{ij}), j=1,...,m},}

\item{U_sm_i}{Discretization points of the ith function: \eqn{U_{i1},\dots,U_{im}}{U_{i1},...,U_{im}}}

\item{K}{Truncation parameter}

\item{pre_smooth}{If pre_smooth==TRUE:  Pre-smoothing of the 'observed' part.  (Reconstruction operator: \eqn{L^*}{L*}). If pre_smooth==FALSE (default): FPCA-estimation of the 'observed' part (Reconstruction operator: \eqn{L}{L})}
}
\description{
This function allows you to reconstruct the missing parts of a function given the observed parts.
}
