% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulFun.R
\name{simuldataKraus}
\alias{simuldataKraus}
\title{Simulate Data as in Kraus (JRSSB, 2015)}
\usage{
simuldataKraus(p = 100, n = 100, a = 0, b = 1)
}
\arguments{
\item{p}{Number of grid points in [a,b]}

\item{n}{Number of functions}

\item{a}{Lower interval boundary}

\item{b}{Upper interval boundary}
}
\description{
This function allows to simulate functional data as used in the simulation study of: 
Kraus, D. (2015). Components and completion of partially observed functional data. 
Journal of the Royal Statistical Society: Series B (Statistical Methodology), 77(4), 777-801. 
However, the missingness process is adjusted to avoid missing 'holes'.
}
\examples{
 
a <- 0; b <- 10; p <- 51; n <- 100
SimDat   <- simuldataKraus(p = p, n = n, a = a, b = b)
## 
Y_mat    <- SimDat[['Y_mat']]
##
matplot(Y_mat[,1:10], col=gray(.5), type="l")
}
