% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstFun.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct partially observed functions}
\usage{
reconstruct(Ly, Lu, K = NULL, K_max = 4, pre_smooth = FALSE,
  nRegGrid = 101)
}
\arguments{
\item{Ly}{List of Y-values. The ith list-element contains \eqn{Y_{i1},\dots,Y_{im}}{Y_{i1},...,Y_{im}}}

\item{Lu}{List of U-values. The ith list-element contains \eqn{U_{i1},\dots,U_{im}}{U_{i1},...,U_{im}}}

\item{K}{Truncation parameter. If K=NULL (default), K is determined using an AIC-type criterion.}

\item{K_max}{Maximum K (used in the AIC-type criterion)}

\item{pre_smooth}{If pre_smooth==TRUE:  Pre-smoothing of the 'observed' part.  (Reconstruction operator: \eqn{L^*}{L*}). If pre_smooth==FALSE (default): FPCA-estimation of the 'observed' part (Reconstruction operator: \eqn{L}{L})}

\item{nRegGrid}{Number of grid-points in the approxmative regular grid.}
}
\description{
This function allows you to reconstruct the missing parts of a function given the observed parts.
}
