% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrausFun.R
\name{reconstructKraus}
\alias{reconstructKraus}
\title{Reconstruct functional data as proposed by Kraus (JRSSB, 2015)}
\usage{
reconstructKraus(X_mat, alpha = NULL)
}
\arguments{
\item{X_mat}{pxn matrix (p: number of discretization points, n=number of functions}

\item{alpha}{Ridge parameter. If alpha = NULL (default), an optimal alpha is determined by GCV}
}
\description{
This function allows to reconstruct partially observed functional data as proposed in: 
Kraus, D. (2015). Components and completion of partially observed functional data. 
Journal of the Royal Statistical Society: Series B (Statistical Methodology), 77(4), 777-801.
}
\examples{
 
a <- 0; b <- 10; p <- 51; n <- 100
SimDat   <- simuldataKraus(p = p, n = n, a = a, b = b)
## 
Y_mat <- SimDat[['Y_mat']]
##
reconst_mat <- reconstructKraus(Y_mat)
##
par(mfrow=c(2,1))
matplot(Y_mat[,1:10],       col=gray(.5), type="l")
matplot(reconst_mat[,1:10], col=gray(.5), type="l")
par(mfrow=c(1,1))
}
